class division mappings

+++ app/Http/Controllers/Institutions/Classifications/ClassDivisionController.php
@@ -5,11 +5,16 @@ use App\Http\Controllers\Controller;
 use App\Models\Institution;
 use App\Models\ClassDivision;
 
+use App\Models\Classification;
+use App\Rules\ValidateExistsRule;
+
 class ClassDivisionController extends Controller
 {
   public function index(Institution $institution)
   {
-    $query = ClassDivision::query();
+    $query = ClassDivision::query()
+      ->withCount('classifications')
+      ->with('classifications');
     return inertia('institutions/class-divisions/list-class-divisions', [
       'classdivisions' => paginateFromRequest($query)
     ]);
@@ -21,13 +26,6 @@ class ClassDivisionController extends Controller
 
     $institution->classDivisions()->create($data);
     return $this->ok();
-  }
-
-  function edit(Institution $institution, ClassDivision $classDivision)
-  {
-    return inertia('institutions/class-divisions/create-edit-class-divisions', [
-      'classDivision' => $classDivision
-    ]);
   }
 
   public function update(Institution $institution, ClassDivision $classDivision)
@@ -51,15 +49,41 @@ class ClassDivisionController extends Controller
     ]);
   }
 
+  public function storeClassification(
+    Institution $institution,
+    ClassDivision $classDivision
+  ) {
+    request()->validate([
+      'classification_id' => [
+        'required',
+        new ValidateExistsRule(Classification::class)
+      ]
+    ]);
+
+    $classDivision->classifications()->attach(request('classification_id'));
+
+    return $this->ok();
+  }
+
+  public function destroyClassification(
+    Institution $institution,
+    ClassDivision $classDivision,
+    Classification $classification
+  ) {
+    $classDivision->classifications()->detach($classification->id);
+
+    return $this->ok();
+  }
+
   public function destroy(
     Institution $institution,
     ClassDivision $classDivision
   ) {
-    // abort_if(
-    //   $classDivision->classifications()->count() > 0,
-    //   403,
-    //   'This class division contains some classes'
-    // );
+    abort_if(
+      $classDivision->classifications()->count() > 0,
+      403,
+      'This class division contains some classes'
+    );
     $classDivision->delete();
     return $this->ok();
   }
